requirements.txt
google-generativeai
tqdm
python-dotenv